terraform/shared/apt-repo/indexer/requirements.txt
boto3
python-debian
python-gnupg
